feature(items_list): replace interval with Update items feature
For current needs, we only need to be able to update items that are selected in the table.
Creating interval will come in future updates.

Created "partials" folder to store the form.

Created an empty `UpdateItemsForm` to avoid confusion with `ScrapeForm` and a new `update_items` view.

Created a `toggleAllCheckboxes` JS script to toggle all checkboxes in the table.

Added a test for `update_items` view