fix(payment_utils): fix token generation algorythm
Added conversion of bool values to the lowercase strings. This is necessary because the API expects boolean values to be lowercase, which lead to incorrect token as a result.
It is now fixed.

Also added a management command `python manage.py get_token` to create a token using test hard-coded info